assistant/pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi[standard]>=0.115.12",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.2",
    "google-auth-oauthlib>=1.2.2",
    "pydantic-ai[logfire]>=0.2.6",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
